+++ src/tests/unit-tests/signal-processing-blocks/tracking/tracking_loop_filter_test.cc
@@ -6,31 +6,21 @@
  *
  * -------------------------------------------------------------------------
  *
- * Copyright (C) 2010-2018  (see AUTHORS file for a list of contributors)
+ * Copyright (C) 2010-2019  (see AUTHORS file for a list of contributors)
  *
  * GNSS-SDR is a software defined Global Navigation
  *          Satellite Systems receiver
  *
  * This file is part of GNSS-SDR.
  *
- * GNSS-SDR is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * GNSS-SDR is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with GNSS-SDR. If not, see <https://www.gnu.org/licenses/>.
+ * SPDX-License-Identifier: GPL-3.0-or-later
  *
  * -------------------------------------------------------------------------
  */
 
 #include "tracking_loop_filter.h"
 #include <gtest/gtest.h>
+#include <cstddef>
 
 TEST(TrackingLoopFilterTest, FirstOrderLoop)
 {
@@ -87,7 +77,7 @@ TEST(TrackingLoopFilterTest, FirstOrderLoopWithLastIntegrator)
     theFilter.initialize(0.0);
 
     float result = 0.0;
-    for (unsigned int i = 0; i < sample_data.size(); ++i)
+    for (size_t i = 0; i < sample_data.size(); ++i)
         {
             result = theFilter.apply(sample_data[i]);
             EXPECT_NEAR(result, expected_out[i], 1e-4);
@@ -118,7 +108,7 @@ TEST(TrackingLoopFilterTest, SecondOrderLoop)
     theFilter.initialize(0.0);
 
     float result = 0.0;
-    for (unsigned int i = 0; i < sample_data.size(); ++i)
+    for (size_t i = 0; i < sample_data.size(); ++i)
         {
             result = theFilter.apply(sample_data[i]);
             EXPECT_NEAR(result, expected_out[i], 1e-4);
@@ -149,7 +139,7 @@ TEST(TrackingLoopFilterTest, SecondOrderLoopWithLastIntegrator)
     theFilter.initialize(0.0);
 
     float result = 0.0;
-    for (unsigned int i = 0; i < sample_data.size(); ++i)
+    for (size_t i = 0; i < sample_data.size(); ++i)
         {
             result = theFilter.apply(sample_data[i]);
             EXPECT_NEAR(result, expected_out[i], 1e-4);
@@ -180,7 +170,7 @@ TEST(TrackingLoopFilterTest, ThirdOrderLoop)
     theFilter.initialize(0.0);
 
     float result = 0.0;
-    for (unsigned int i = 0; i < sample_data.size(); ++i)
+    for (size_t i = 0; i < sample_data.size(); ++i)
         {
             result = theFilter.apply(sample_data[i]);
             EXPECT_NEAR(result, expected_out[i], 1e-4);
@@ -211,7 +201,7 @@ TEST(TrackingLoopFilterTest, ThirdOrderLoopWithLastIntegrator)
     theFilter.initialize(0.0);
 
     float result = 0.0;
-    for (unsigned int i = 0; i < sample_data.size(); ++i)
+    for (size_t i = 0; i < sample_data.size(); ++i)
         {
             result = theFilter.apply(sample_data[i]);
             EXPECT_NEAR(result, expected_out[i], 1e-4);
